gatewaytest: refactor out a common testing package
