+++ go/src/koding/kites/gateway/gateway_test.go
@@ -7,48 +7,25 @@ import (
 	"time"
 
 	"koding/kites/gateway"
+	"koding/kites/gateway/gatewaytest"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/s3"
-	"github.com/koding/logging"
 )
 
-type Flags struct {
-	AccessKey string        `required:"true"`
-	SecretKey string        `required:"true"`
-	Bucket    string        `default:"kodingdev-publiclogs"`
-	Region    string        `default:"us-east-1"`
-	Expire    time.Duration `default:"15m0s"`
-}
-
 func TestGateway_UserBucket(t *testing.T) {
-	var f Flags
+	var f gatewaytest.Flags
 
-	if err := ParseFlags(&f); err != nil {
+	if err := gatewaytest.ParseFlags(&f); err != nil {
 		t.Fatal(err)
 	}
 
 	// to cleanup after tests
-	rootS3 := s3.New(session.New(&aws.Config{
-		Credentials: credentials.NewStaticCredentials(f.AccessKey, f.SecretKey, ""),
-		Region:      &f.Region,
-	}))
-
-	drv := &Driver{
-		ChanCap: 1,
-	}
-
-	cfg := &gateway.Config{
-		AccessKey:  f.AccessKey,
-		SecretKey:  f.SecretKey,
-		Bucket:     f.Bucket,
-		AuthExpire: f.Expire,
-		Region:     f.Region,
-		Log:        logging.NewCustom("gateway-test", testing.Verbose()),
-	}
+	rootS3 := s3.New(session.New(f.AWSConfig()))
+	drv := &gatewaytest.Driver{ChanCap: 1}
+	cfg := f.Config()
 
 	defer drv.Server(cfg)()
